Move checkMemberMeetsCategory to coffeeChatPAI
