+++ backend/src/API/memberAPI.ts
@@ -3,7 +3,6 @@ import PermissionsManager from '../utils/permissionsManager';
 import { BadRequestError, PermissionError } from '../utils/errors';
 import { getNetIDFromEmail, computeMembersDiff } from '../utils/memberUtil';
 import { deleteImage } from './imageAPI';
-import { getMemberProperties } from './coffeeChatAPI';
 
 const membersDao = new MembersDao();
 
@@ -169,83 +168,4 @@ export const generateMemberArchive = async (
     );
   }
   return MembersDao.generateArchive(membershipChanges, semesters);
-};
-
-/**
- * Checks if a member meets a category.
- * @param otherMemberEmail - the email of the member we are checking.
- * @param submitter - the member that submitted the coffee chat.
- * @param encodedCategory - the category we are checking with (encoded with base64).
- * @returns true if a member meets a category, false if not, undefined if not enough data to know.
- */
-export const checkMemberMeetsCategory = async (
-  otherMemberEmail: string,
-  submitter: IdolMember,
-  encodedCategory: string
-): Promise<boolean | undefined> => {
-  const otherMemberProperties = await getMemberProperties(otherMemberEmail);
-  const submitterProperties = await getMemberProperties(submitter.email);
-  const otherMember = await getMember(otherMemberEmail);
-  const decode = (input: string): string => input.replace(/\*/g, '/');
-  const category = decode(encodedCategory);
-  const haveNoCommonSubteams = (member1: IdolMember, member2: IdolMember): boolean =>
-    member2.subteams.every((team) => !member1.subteams.includes(team)) &&
-    member1.subteams.every((team) => !member2.subteams.includes(team));
-
-  if (category === 'an alumni') {
-    return (await allMembers()).every((member) => member.email !== otherMember?.email);
-  }
-  if (category === 'courseplan member') {
-    return otherMember?.subteams.includes('courseplan');
-  }
-  if (category === 'a pm (not your team)') {
-    return otherMember?.role === 'pm' && haveNoCommonSubteams(submitter, otherMember);
-  }
-  if (category === 'business member') {
-    return otherMember?.role === 'business';
-  }
-  if (category === 'is/was a TA') {
-    return otherMemberProperties ? otherMemberProperties.ta : undefined;
-  }
-  if (category === 'major/minor that is not cs/infosci') {
-    return otherMemberProperties ? otherMemberProperties.notCsOrInfosci : undefined;
-  }
-  if (category === 'idol member') {
-    return otherMember?.subteams.includes('idol');
-  }
-  if (category === 'a newbie') {
-    return otherMemberProperties ? otherMemberProperties.newbie : undefined;
-  }
-  if (category === 'from a different college') {
-    return otherMemberProperties && submitterProperties
-      ? otherMemberProperties.college !== submitterProperties.college
-      : undefined;
-  }
-  if (category === 'curaise member') {
-    return otherMember?.subteams.includes('curaise');
-  }
-  if (category === 'cornellgo member') {
-    return otherMember?.subteams.includes('cornellgo');
-  }
-  if (category === 'a tpm (not your team)') {
-    return otherMember?.role === 'tpm' && haveNoCommonSubteams(submitter, otherMember);
-  }
-  if (category === 'carriage member') {
-    return otherMember?.subteams.includes('carriage');
-  }
-  if (category === 'qmi member') {
-    return otherMember?.subteams.includes('queuemein');
-  }
-  if (category === 'a lead (not your role)') {
-    return (
-      otherMember?.role === 'lead' &&
-      (submitter.role !== 'lead'
-        ? otherMemberProperties?.leadType !== submitter.role
-        : otherMemberProperties?.leadType !== submitterProperties?.leadType)
-    );
-  }
-  if (category === 'cuapts member') {
-    return otherMember?.subteams.includes('cuapts');
-  }
-  return undefined;
 };
