+++ frontend/src/components/Forms/CoffeeChatsForm/CoffeeChatsForm.tsx
@@ -32,7 +32,7 @@ const CoffeeChatsForm: React.FC = () => {
     });
     CoffeeChatAPI.getCoffeeChatBingoBoard().then((board) => setBingoBoard(board));
     if (member && category) {
-      MembersAPI.checkMemberMeetsCategory(member, category).then((check) =>
+      CoffeeChatAPI.checkMemberMeetsCategory(member, category).then((check) =>
         setMemberMeetsCategory(check)
       );
     }
